Refactor converters: simple (no norm conversion, data-norm, spike-norm, few-spike are chosen with --converter command line option.
