+++ ml_genn/converters/__init__.py
@@ -1,2 +1,5 @@
+from ml_genn.converters.enum import ConverterType
+from ml_genn.converters.simple import Simple
 from ml_genn.converters.few_spike import FewSpike
-from ml_genn.converters.rate_based import RateBased
+from ml_genn.converters.data_norm import DataNorm
+from ml_genn.converters.spike_norm import SpikeNorm
