+++ ml_genn/layers/enum.py
@@ -1,5 +1,12 @@
 from enum import Enum
 
+class InputType(Enum):
+    SPIKE = 'spike'
+    SPIKE_SIGNED = 'spike_signed'
+    POISSON = 'poisson'
+    POISSON_SIGNED = 'poisson_signed'
+    IF = 'if'
+
 class ConnectivityType(Enum):
     PROCEDURAL = 'procedural'
     SPARSE = 'sparse'
