+++ ml_genn/model.py
@@ -23,7 +23,7 @@ import tensorflow as tf
 from tqdm import tqdm
 from pygenn.genn_model import GeNNModel
 
-from ml_genn.converters import RateBased
+from ml_genn.converters import Simple
 from ml_genn.layers import InputLayer
 from ml_genn.layers import Dense
 from ml_genn.layers import AvePool2DDense
@@ -259,7 +259,7 @@ class Model(object):
 
 
     @staticmethod
-    def convert_tf_model(tf_model, converter=RateBased('poisson'),
+    def convert_tf_model(tf_model, converter=Simple(),
                          connectivity_type='procedural', **compile_kwargs):
         """Create a ML GeNN model from a TensorFlow model
 
@@ -289,14 +289,14 @@ class Model(object):
             elif isinstance(tf_layer, (tf.keras.layers.Dense, tf.keras.layers.Conv2D)):
                 converter.validate_tf_layer(tf_layer)
 
-        # Allow converter to perform any pre-compilation normalisation
-        norm_output = converter.normalise_pre_compile(tf_model)
-        
+        # Perform any pre-compilation tasks
+        pre_compile_output = converter.pre_compile(tf_model)
+
         model = Model(name=tf_model.name)
 
         # Add input layer
         layer = InputLayer('input', tf_model.input_shape[1:], 
-                           converter.create_input_neurons(norm_output))
+                           converter.create_input_layer(pre_compile_output))
         model.inputs.append(layer)
         model.layers.append(layer)
         previous_layer = layer
@@ -317,8 +317,7 @@ class Model(object):
                 if pool_layer is None:
                     print('converting Dense layer <{}>'.format(tf_layer.name))
                     layer = Dense(name=tf_layer.name, units=tf_layer.units,
-                                  neurons=converter.create_neurons(tf_layer,
-                                                                   norm_output))
+                                  neurons=converter.create_layer(tf_layer, pre_compile_output))
                 else:
                     print('converting AveragePooling2D -> Dense layers <{}>'.format(tf_layer.name))
                     layer = AvePool2DDense(
@@ -327,8 +326,7 @@ class Model(object):
                         pool_strides=pool_layer.strides,
                         pool_padding=pool_layer.padding,
                         connectivity_type=connectivity_type, 
-                        neurons=converter.create_neurons(tf_layer, 
-                                                         norm_output))
+                        neurons=converter.create_layer(tf_layer, pre_compile_output))
 
                 layer.connect([previous_layer])
                 layer.set_weights(tf_layer.get_weights())
@@ -347,8 +345,7 @@ class Model(object):
                         conv_strides=tf_layer.strides,
                         conv_padding=tf_layer.padding,
                         connectivity_type=connectivity_type, 
-                        neurons=converter.create_neurons(tf_layer, 
-                                                         norm_output))
+                        neurons=converter.create_layer(tf_layer, pre_compile_output))
                 else:
                     print('converting AveragePooling2D -> Conv2D layers <{}>'.format(tf_layer.name))
                     layer = AvePool2DConv2D(
@@ -357,8 +354,7 @@ class Model(object):
                         pool_strides=pool_layer.strides, conv_strides=tf_layer.strides,
                         pool_padding=pool_layer.padding, conv_padding=tf_layer.padding,
                         connectivity_type=connectivity_type, 
-                        neurons=converter.create_neurons(tf_layer,
-                                                         norm_output))
+                        neurons=converter.create_layer(tf_layer, pre_compile_output))
 
                 layer.connect([previous_layer])
                 layer.set_weights(tf_layer.get_weights())
@@ -378,6 +374,7 @@ class Model(object):
         # Compile model
         model.compile(**compile_kwargs)
         
-        # Perform any post-compilation normalisation operations that might be required
-        converter.normalise_post_compile(tf_model, model)
+        # Perform any post-compilation tasks
+        converter.post_compile(model)
+
         return model
