+++ tests/test_avepool2d_dense.py
@@ -7,7 +7,7 @@ def model_compare_tf_and_mlg(tf_model, x, connectivity_type='procedural'):
     tf_y = tf_model(x).numpy()
 
     # Run ML GeNN model
-    mlg_model = mlg.Model.convert_tf_model(tf_model, converter=mlg.converters.RateBased('spike'), 
+    mlg_model = mlg.Model.convert_tf_model(tf_model, converter=mlg.converters.Simple('spike'), 
                                            connectivity_type=connectivity_type,
                                            dt=1.0, batch_size=1)
